+++ src/main/java/br/com/zup/gerenciador/de/dailys/impedimento/ImpedimentoService.java
@@ -1,7 +1,7 @@
-package br.com.zup.gerenciador.de.dailys.service;
+package br.com.zup.gerenciador.de.dailys.impedimento;
 
-import br.com.zup.gerenciador.de.dailys.model.Impedimento;
-import br.com.zup.gerenciador.de.dailys.repository.ImpedimentoRepository;
+import br.com.zup.gerenciador.de.dailys.impedimento.Impedimento;
+import br.com.zup.gerenciador.de.dailys.impedimento.ImpedimentoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
