+++ src/main/java/br/com/zup/gerenciador/de/dailys/taskAtual/TaskAtual.java
@@ -1,14 +1,12 @@
-package br.com.zup.gerenciador.de.dailys.model;
+package br.com.zup.gerenciador.de.dailys.taskAtual;
 
+import br.com.zup.gerenciador.de.dailys.usuario.Usuario;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import javax.persistence.*;
 import java.util.Date;
 
 @Entity
@@ -17,14 +15,17 @@ import java.util.Date;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class Impedimento {
+public class TaskAtual {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String descricao;
+    private Date dataInicio;
+    private Date previsaoFim;
+
+    @ManyToOne(optional = false)
+    private Usuario usuarios;
+
 
 }
-
-
-
