+++ src/main/java/br/com/zup/gerenciador/de/dailys/usuario/UsuarioController.java
@@ -1,7 +1,7 @@
-package br.com.zup.gerenciador.de.dailys.controller;
+package br.com.zup.gerenciador.de.dailys.usuario;
 
 
-import br.com.zup.gerenciador.de.dailys.service.UsuarioService;
+import br.com.zup.gerenciador.de.dailys.usuario.UsuarioService;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
