+++ src/main/java/br/com/zup/gerenciador/de/dailys/usuario/UsuarioRepository.java
@@ -1,6 +1,5 @@
-package br.com.zup.gerenciador.de.dailys.repository;
+package br.com.zup.gerenciador.de.dailys.usuario;
 
-import br.com.zup.gerenciador.de.dailys.model.Usuario;
 import org.springframework.data.repository.CrudRepository;
 
 public interface UsuarioRepository extends CrudRepository<Usuario, String> {
